Fixed erros in Oracle and ArrConversion contract

+++ contracts/Helper/ArrConversion.sol
@@ -10,40 +10,75 @@ contract ArrConversion {
         return arr(uint256(_a));
     }
 
-    function arr(bytes32 _a, bytes32 _b) internal pure returns (uint256[] memory r) {
+    function arr(bytes32 _a, bytes32 _b)
+        internal
+        pure
+        returns (uint256[] memory r)
+    {
         return arr(uint256(_a), uint256(_b));
     }
 
     function arr(address _a) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a));
+        return arr(uint160(_a));
     }
 
-    function arr(address _a, address _b) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), uint256(_b));
+    function arr(address _a, address _b)
+        internal
+        pure
+        returns (uint256[] memory r)
+    {
+        return arr(uint160(_a), uint160(_b));
     }
 
-    function arr(address _a, uint256 _b, uint256 _c) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), _b, _c);
+    function arr(
+        address _a,
+        uint256 _b,
+        uint256 _c
+    ) internal pure returns (uint256[] memory r) {
+        return arr(uint160(_a), _b, _c);
     }
 
-    function arr(address _a, uint256 _b, uint256 _c, uint256 _d) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), _b, _c, _d);
+    function arr(
+        address _a,
+        uint256 _b,
+        uint256 _c,
+        uint256 _d
+    ) internal pure returns (uint256[] memory r) {
+        return arr(uint160(_a), _b, _c, _d);
     }
 
-    function arr(address _a, uint256 _b) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), uint256(_b));
+    function arr(address _a, uint256 _b)
+        internal
+        pure
+        returns (uint256[] memory r)
+    {
+        return arr(uint160(_a), uint160(_b));
     }
 
-    function arr(address _a, address _b, uint256 _c, uint256 _d, uint256 _e) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), uint256(_b), _c, _d, _e);
+    function arr(
+        address _a,
+        address _b,
+        uint256 _c,
+        uint256 _d,
+        uint256 _e
+    ) internal pure returns (uint256[] memory r) {
+        return arr(uint160(_a), uint160(_b), _c, _d, _e);
     }
 
-    function arr(address _a, address _b, address _c) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), uint256(_b), uint256(_c));
+    function arr(
+        address _a,
+        address _b,
+        address _c
+    ) internal pure returns (uint256[] memory r) {
+        return arr(uint160(_a), uint160(_b), uint160(_c));
     }
 
-    function arr(address _a, address _b, uint256 _c) internal pure returns (uint256[] memory r) {
-        return arr(uint256(_a), uint256(_b), uint256(_c));
+    function arr(
+        address _a,
+        address _b,
+        uint256 _c
+    ) internal pure returns (uint256[] memory r) {
+        return arr(uint160(_a), uint160(_b), uint160(_c));
     }
 
     function arr(uint256 _a) internal pure returns (uint256[] memory r) {
@@ -51,20 +86,33 @@ contract ArrConversion {
         r[0] = _a;
     }
 
-    function arr(uint256 _a, uint256 _b) internal pure returns (uint256[] memory r) {
+    function arr(uint256 _a, uint256 _b)
+        internal
+        pure
+        returns (uint256[] memory r)
+    {
         r = new uint256[](2);
         r[0] = _a;
         r[1] = _b;
     }
 
-    function arr(uint256 _a, uint256 _b, uint256 _c) internal pure returns (uint256[] memory r) {
+    function arr(
+        uint256 _a,
+        uint256 _b,
+        uint256 _c
+    ) internal pure returns (uint256[] memory r) {
         r = new uint256[](3);
         r[0] = _a;
         r[1] = _b;
         r[2] = _c;
     }
 
-    function arr(uint256 _a, uint256 _b, uint256 _c, uint256 _d) internal pure returns (uint256[] memory r) {
+    function arr(
+        uint256 _a,
+        uint256 _b,
+        uint256 _c,
+        uint256 _d
+    ) internal pure returns (uint256[] memory r) {
         r = new uint256[](4);
         r[0] = _a;
         r[1] = _b;
@@ -72,7 +120,13 @@ contract ArrConversion {
         r[3] = _d;
     }
 
-    function arr(uint256 _a, uint256 _b, uint256 _c, uint256 _d, uint256 _e) internal pure returns (uint256[] memory r) {
+    function arr(
+        uint256 _a,
+        uint256 _b,
+        uint256 _c,
+        uint256 _d,
+        uint256 _e
+    ) internal pure returns (uint256[] memory r) {
         r = new uint256[](5);
         r[0] = _a;
         r[1] = _b;
